[HAL] Fix and document addressable LED timings
HAL_SetAddressableLEDBitTiming swapped high and low timings for whatever
was written to it. This fixes that bug.

Additionally, the API has been updated to take high time first, and then
low time. This is due to this being the physical data format, so having
the api match is clearer.

Additionally, update the docs with the defaults. Replaces #5271