[wpinet]Leak multicast handles during windows shutdown
Destructing either of the multicast objects during process shutdown will result in a crash due to attempting to start a task on the non existant thread pool.

Solve this by just leaking all the handles upon destruction of the static multicast manager. This won't solve the case where the user statically allocates the object, but solves Java and C access, and most cases wouldn't be statically allocation the service anouncer anyway in C++