Remove static JNI loading
Static JNI loading is impossible to catch, and results in weird error messages

Also, only robot projects use it. All the tools explicitly disable it because they load a different way. So its safe to remove. The only thing it might break is unit tests, we will see. If so, its easy to document the call required.