load window settings without erasing them.
Fixes 6036.

Tested on Ubuntu 22.04, using cmake and gradle

The change is simply to move the WorkspaceResetImpl() before
LoadWindowStorageImpl() instead of after.

It looks to me like this has always been broken, which
doesn't make any sense to me, maybe somebody more familiar
can help figure it out.  Regardless of the former scope of the
issue, it seems good now, which is what I wanted.