Fix stringop overflow warnings from GCC 14
If ConvertToC(size_t) is passed a string with length
std::numeric_limits<size_t>::max(), the wpi::safe_malloc(in.size() + 1)
call will overflow to 0, causing wpi::safe_malloc() to return a buffer
with length 1. This isn't enough to contain the input string.

We now report failure if a string of that size is provided, because it's
impossible to convert to a C-style string.