Add enforcement of wpilib's java style when using vscode.
This will automatically format modified lines (according to wpilib's java style) when a java file is saved in vscode.

This requires including .vscode/settings.json in the repo which appears to be a fairly widely accepted practice in large projects (including vscode itself and mozilla). For details, see: https://stackoverflow.com/questions/32964920/should-i-commit-the-vscode-folder-to-source-control