[ci] Add workflow to build java tools for PRs
This can catch breaking changes earlier
Builds linux only for speed
Builds and publishes artifacts in this file as it was actually faster
then reusing the artifacts from the gradle file as those need the
combiner run on them. It uploads the artifacts to Actions with a 1 day
retention, because that allows the tools to be built in parallel, which
overall sped things up. This also only allows relevant tasks in wpilib
to be run
The tool builds are uploaded with a shorter retention time in case
someone wants to do more extensive tests.