applyDeadband for translation2d and 3d
The same idea as applyDeadband in MathUtil. this is useful in scenarios like applying a deadband for joystick input, where you want the x and y inputs to be zeros when the norm of that translation is sufficiently small. (If use deadbands separately, you encounter the issue of "x value is forced to zero but y is not, thus the swerves suddenly align themselves in the y-direction. According to our driver that is uncomfortable.) I considered about adding this to the Vector class but Translation is more popular i think.