[bazel] Link static and shared libraries correctly
This gets us to the point where most of the static and shared libraries
are built correctly, and match the gradle versions.  I've verified this
with a script which compares artifacts.  The packaging and release parts
still need to be finished, but this is big enough already.

Chunks of this code are from PJ, the rest is from me.  I've lost track
of which pieces are from whom.

cc_shared_library does most everything we want to generate the correct
graph of shared libraries.  I ended up needing to create my own version
of it for static libraries, since we want similar complexity when
splitting things up, and the upstream cc_static_library isn't flexible
enough.

version.bzl is currently manually updated, I'd like to update it
automatically.  Everything which needs a version sources from there.
Ideally, --embed_label would populate it with a repository rule.  Feels
like a good follow-on commit.

Signed-off-by: Austin Schuh <austin.linux@gmail.com>