Run child commands BEFORE parents, not after
This removes a 1-loop cycle penalty for parents to resume after a child completes. The penalty was caused by the parent checking if the child completed /before/ the child ran, so the first time a parent would have been able to resume would be the loop immediately after the child completed.

Actual command ordering is unchanged (eg ordering in telemetry is still [parent, child])