[java] Support disabling Preferences support for legacy dashboards
To support legacy dashboards, the Preferences code installed a listener
that marked all new topics as persisted. Unfortunately, having a
listener that updates new topics in the background can result in a
closed NetworkTablesInstance being accessed (resulting in a SIGSEGV)
when Preferences.setNetworkTableInstance() is called soon after values
have been published for new topics (see #8215).

It's possible this race condition is why PreferencesTest.removeAllTest()
was flaky.

The code to set new topics discovered under /Preferences as persistent
was added back in 2015 in 87fc49c. Dashboards that want to want data for
new keys to be saved should use the Preferences API so the topics are
marked as persistent.